refactor(web): normalize file naming (.constants/.styles) and URL constants
