+++ apps/web/src/app/[locale]/dev/token/page.tsx
@@ -5,7 +5,7 @@ import { Button } from '@gitanimals/ui-tailwind';
 import { useClientSession } from '@/utils/clientAuth';
 import { copyClipBoard } from '@/utils/copy';
 
-import * as styles from './token.style';
+import * as styles from './token.styles';
 
 function DevTokenPage() {
   const { data } = useClientSession();
