+++ apps/web/src/app/[locale]/laboratory/_components/ConfirmDialog.tsx
