+++ apps/web/src/app/[locale]/laboratory/_components/LaboratoryLayout.tsx
