+++ apps/web/src/app/[locale]/landing/MainSection/MainSlider.styles.ts
