+++ apps/web/src/app/[locale]/landing/RankingSection/RankingSection.constants.ts
