+++ apps/web/src/components/AnimalCard/AnimalCard.constants.ts
