+++ apps/web/src/constants/outlink.ts
@@ -1,4 +1,5 @@
 export const STATIC_IMAGE_URL = 'https://static.gitanimals.org';
+export const RENDER_URL = 'https://render.gitanimals.org';
 export const GIT_ANIMALS_MAIN_URL = 'https://github.com/git-goods/gitanimals';
 
 export const SUMI_GITHUB_URL = 'https://github.com/sumi-0011';
